Move comparator into analog
