+++ src/analog/comparator.rs
