+++ src/analog/mod.rs
@@ -1,3 +1,5 @@
 pub mod adc;
 #[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
+pub mod comparator;
+#[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
 pub mod dac;
