+++ src/prelude.rs
@@ -6,15 +6,15 @@ pub use hal::watchdog::WatchdogEnable as _;
 
 pub use crate::analog::adc::AdcExt as _;
 #[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
+pub use crate::analog::comparator::ComparatorExt as _;
+#[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
+pub use crate::analog::comparator::ComparatorSplit as _;
+#[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
+pub use crate::analog::comparator::WindowComparatorExt as _;
+#[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
 pub use crate::analog::dac::DacExt as _;
 #[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
 pub use crate::analog::dac::DacOut as _;
-#[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
-pub use crate::comparator::ComparatorExt as _;
-#[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
-pub use crate::comparator::ComparatorSplit as _;
-#[cfg(any(feature = "stm32g071", feature = "stm32g081"))]
-pub use crate::comparator::WindowComparatorExt as _;
 pub use crate::crc::CrcExt as _;
 pub use crate::delay::DelayExt as _;
 // pub use crate::dma::CopyDma as _;
